+++ library/PkiExtension/PkiExtension/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace PkiExtension
-{
-    public class Class1
-    {
-    }
-}