+++ PkiExtension/changelog.md
@@ -1,5 +1,11 @@
 ﻿# Changelog
 
+## 1.1.3 (2025-09-25)
+
++ New: Get-PkiCaCertificate - Lists certificates from a CA.
++ Upd: Get-PkiCaExpiringCertificate - Added `-Server` parameter to be able to resolve templates from another forest
++ Upd: Get-PkiCaExpiringCertificate - Added `-Properties` parameter to be able to select the properties returned
+
 ## 1.0.0 (???)
 
 + Initial Release